server-api/requirements.txt
Flask
Flask-OIDC
requests
flask_cors
python-keycloak
PyJWT
